softsand.py represents soft sand rock physics model, added 4D gravity and 4D avo
